♻️ Refactor: chat 컴포넌트 폴더 내부의 util, constant 분리
