+++ Frontend/src/constants/chat/color.ts
@@ -1,0 +1,12 @@
+export const TAILWIND_USER_COLORS = [
+  'text-rose-500', // 붉은색
+  'text-orange-500', // 주황색
+  'text-amber-500', // 황색
+  'text-lime-500', // 라임색
+  'text-emerald-500', // 에메랄드색
+  'text-cyan-500', // 시안색
+  'text-blue-500', // 파란색
+  'text-violet-500', // 보라색
+  'text-purple-500', // 자주색
+  'text-pink-500', // 분홍색
+] as const;
