+++ Frontend/src/utils/chatUtils.ts
@@ -1,4 +1,4 @@
-import { TAILWIND_USER_COLORS } from './constants';
+import { TAILWIND_USER_COLORS } from '@constants/chat/color';
 
 // 텍스트의 각 문자 코드 합을 기반으로 일관된 테일윈드 스타일 색상을 반환하는 함수
 export const getConsistentTextColor = (text: string): (typeof TAILWIND_USER_COLORS)[number] =>
